refactor: 프로필에 표시될 자유 게시물 관련 로직 분리
