+++ src/main/java/com/trendist/post_service/domain/post/controller/PostController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 import com.trendist.post_service.domain.post.dto.request.PostUpdateRequest;
 import com.trendist.post_service.domain.post.dto.response.PostDeleteResponse;
 import com.trendist.post_service.domain.post.dto.response.PostGetAllResponse;
-import com.trendist.post_service.domain.post.dto.response.PostGetMineResponse;
 import com.trendist.post_service.domain.post.dto.response.PostGetResponse;
 import com.trendist.post_service.domain.post.dto.response.PostLikeResponse;
 import com.trendist.post_service.domain.post.dto.response.PostUpdateResponse;
@@ -87,15 +86,6 @@ public class PostController {
 	@GetMapping("/{postId}")
 	public ApiResponse<PostGetResponse> getPost(@PathVariable(name = "postId") UUID postId) {
 		return ApiResponse.onSuccess(postService.getPost(postId));
-	}
-
-	@Operation(
-		summary = "내가 쓴 자유 게시판 게시물 조회",
-		description = "현재 로그인한 사용자가 자유 게시판에 자신이 생성한 게시물들을 조회합니다."
-	)
-	@GetMapping("/mine")
-	public ApiResponse<Page<PostGetMineResponse>> getMyPosts(@RequestParam(defaultValue = "0") int page) {
-		return ApiResponse.onSuccess(postService.getMyPosts(page));
 	}
 
 	@Operation(
